requirements.txt
streamlit>=1.32.0
groq>=1.0.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
